+++ montepy/input_parser/mcnp_input.py
@@ -69,9 +69,6 @@ class ParsingNode(ABC):
     """
     Object to represent a single coherent MCNP input, such as an input.
 
-    .. versionadded:: 0.2.0
-        This was added as part of the parser rework.
-
     :param input_lines: the lines read straight from the input file.
     :type input_lines: list
     """
@@ -113,10 +110,6 @@ class ParsingNode(ABC):
 class Input(ParsingNode):
     """
     Represents a single MCNP "Input" e.g. a single cell definition.
-
-    .. versionadded:: 0.2.0
-        This was added as part of the parser rework, and rename.
-        This was a replacement for :class:`Card`.
 
     :param input_lines: the lines read straight from the input file.
     :type input_lines: list