+++ montepy/input_parser/surface_parser.py
@@ -8,9 +8,6 @@ class SurfaceParser(MCNP_Parser):
     """
     A parser for MCNP surfaces.
 
-    .. versionadded:: 0.2.0
-        This was added with the major parser rework.
-
     :rtype: SyntaxNode
     """