+++ montepy/mcnp_problem.py
@@ -19,7 +19,7 @@ from montepy.surface_collection import Surfaces
 from montepy.data_inputs import parse_data
 from montepy.input_parser import input_syntax_reader, block_type, mcnp_input
 from montepy.input_parser.input_file import MCNP_InputFile
-from montepy.universes import Universes
+from montepy.universes import Universe, Universes
 from montepy.transforms import Transforms
 import montepy
 
@@ -40,7 +40,7 @@ class MCNP_Problem:
         surface.Surface: Surfaces,
         Material: Materials,
         transform.Transform: Transforms,
-        montepy.universe.Universe: Universes,
+        Universe: Universes,
     }
 
     def __init__(self, destination):
@@ -339,7 +339,7 @@ class MCNP_Problem:
         OBJ_MATCHER = {
             block_type.BlockType.CELL: (Cell, self._cells),
             block_type.BlockType.SURFACE: (
-                surface_builder.surface_builder,
+                surface_builder.parse_surface,
                 self._surfaces,
             ),
             block_type.BlockType.DATA: (parse_data, self._data_inputs),
@@ -623,3 +623,60 @@ class MCNP_Problem:
                 ret += f"{obj}\n"
             ret += "\n"
         return ret
+
+    def parse(self, input: str, append: bool = True) -> montepy.mcnp_object.MCNP_Object:
+        """
+        Parses the MCNP object given by the string, and links it adds it to this problem.
+
+        This attempts to identify the input type by trying to parse it in the following order:
+
+        #. Data Input
+        #. Surface
+        #. Cell
+
+        This is done mostly for optimization to go from easiest parsing to hardest.
+        This will:
+
+        #. Parse the input
+        #. Link it to other objects in the problem. Note: this will raise an error if those objects don't exist.
+        #. Append it to the appropriate collection
+
+        :param input: the string describing the input. New lines are allowed but this does not need to meet MCNP line
+            length rules.
+        :type input: str
+        :param append: Whether to append this parsed object to this problem.
+        :type append: bool
+        :returns: the parsed object.
+        :rtype: MCNP_Object
+
+        :raises TypeError: If a str is not given
+        :raises ParsingError: If this is not a valid input.
+        :raises BrokenObjectLinkError: if the dependent objects are not already in the problem.
+        :raises NumberConflictError: if the object's number is already taken
+        """
+        try:
+            obj = montepy.parse_data(input)
+        except ParsingError:
+            try:
+                obj = montepy.parse_surface(input)
+            except ParsingError:
+                obj = montepy.Cell(input)
+                # let final parsing error bubble up
+        obj.link_to_problem(self)
+        if isinstance(obj, montepy.Cell):
+            obj.update_pointers(self.cells, self.materials, self.surfaces)
+            if append:
+                self.cells.append(obj)
+        elif isinstance(obj, montepy.surfaces.surface.Surface):
+            obj.update_pointers(self.surfaces, self.data_inputs)
+            if append:
+                self.surfaces.append(obj)
+        else:
+            obj.update_pointers(self.data_inputs)
+            if append:
+                self.data_inputs.append(obj)
+                if isinstance(obj, Material):
+                    self._materials.append(obj, insert_in_data=False)
+                if isinstance(obj, transform.Transform):
+                    self._transforms.append(obj, insert_in_data=False)
+        return obj